hw2/vacation_agency/requirements.txt
amqp==5.1.1
appnope==0.1.3
asgiref==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.6
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.1
Deprecated==1.2.13
Django==3.2.13
django-celery-beat==2.2.1
django-celery-results==2.3.1
django-extensions==3.1.5
django-filter==21.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
executing==0.8.3
flake8==4.0.1
importlib-metadata==4.11.3
install==1.3.5
ipython==8.2.0
jedi==0.18.1
kombu==5.2.4
Markdown==3.3.6
matplotlib-inline==0.1.3
mccabe==0.6.1
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.11.2
pyparsing==3.0.8
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
redis==4.2.2
roman==3.3
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.1.1
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.1
wrapt==1.14.1
zipp==3.8.0
